Fix end-points
