+++ src/main/webapp/js/client.js
@@ -1,7 +1,4 @@
-// const loadingElement = document.querySelector(".loading");
 const dataElement = document.querySelector(".data");
-
-// loadingElement.style.display = "none";
 
 function listAllMatches() {
     getMatch().then((matches) => {
@@ -33,57 +30,55 @@ function listAllMatches() {
 
             dataElement.appendChild(div);
         });
-    });
-}
 
 function listAllCharities() {
-    getCharities().then((charities) => {
-        dataElement.innerHTML = "";
-        charities.forEach((charity) => {
-            const div = document.createElement("div");
+  getCharities().then((charities) => {
+    dataElement.innerHTML = "";
+    charities.forEach((charity) => {
+      const div = document.createElement("div");
 
-            const header = document.createElement("h3");
-            header.textContent = charity.name;
+      const header = document.createElement("h3");
+      header.textContent = charity.CharityName;
 
-            const contents = document.createElement("p");
-            contents.textContent = charity.description;
+      const description = document.createElement("p");
+      description.textContent = charity.description;
 
-            const address = document.createElement("p");
-            address.textContent = charity.address;
+      const location = document.createElement("p");
+      location.textContent = charity.location;
 
-            div.appendChild(header);
-            div.appendChild(contents);
-            div.appendChild(address);
+      div.appendChild(header);
+      div.appendChild(description);
+      div.appendChild(location);
 
-            dataElement.appendChild(div);
-        });
-        loadingElement.style.display = "none";
+      dataElement.appendChild(div);
     });
+    loadingElement.style.display = "none";
+  });
 }
 
 function listAllRestaurants() {
-    getRestaurants().then((restaurants) => {
-        dataElement.innerHTML = "";
-        restaurants.forEach((restaurant) => {
-            const div = document.createElement("div");
+  getRestaurants().then((restaurants) => {
+    dataElement.innerHTML = "";
+    restaurants.forEach((restaurant) => {
+      const div = document.createElement("div");
 
-            const header = document.createElement("h3");
-            header.textContent = restaurant.name;
+      const header = document.createElement("h3");
+      header.textContent = restaurant.restaurantName;
 
-            const contents = document.createElement("p");
-            contents.textContent = restaurant.description;
+      const description = document.createElement("p");
+      description.textContent = restaurant.description;
 
-            const address = document.createElement("p");
-            address.textContent = restaurant.address;
+      const location = document.createElement("p");
+      location.textContent = restaurant.location;
 
-            div.appendChild(header);
-            div.appendChild(contents);
-            div.appendChild(address);
+      div.appendChild(header);
+      div.appendChild(description);
+      div.appendChild(location);
 
-            dataElement.appendChild(div);
-        });
-        loadingElement.style.display = "none";
+      dataElement.appendChild(div);
     });
+    loadingElement.style.display = "none";
+  });
 }
 
 function getMatch() {
@@ -95,47 +90,19 @@ function getMatch() {
 }
 
 function getCharities() {
-    const charities = [{
-            name: "GiveDirectly",
-            description: "GiveDirectly is the first — and largest — nonprofit that lets donors like you send food directly to the world’s poorest. We believe people living in poverty deserve the dignity to choose for themselves how best to improve their lives.",
-            address: "Kent Ridge | Singapore",
-        },
-        {
-            name: "GiveDirectly",
-            description: "GiveDirectly is the first — and largest — nonprofit that lets donors like you send food directly to the world’s poorest. We believe people living in poverty deserve the dignity to choose for themselves how best to improve their lives.",
-            address: "Kent Ridge | Singapore",
-        },
-        {
-            name: "GiveDirectly",
-            description: "GiveDirectly is the first — and largest — nonprofit that lets donors like you send food directly to the world’s poorest. We believe people living in poverty deserve the dignity to choose for themselves how best to improve their lives.",
-            address: "Kent Ridge | Singapore",
-        },
-    ];
-
-    // Using Promise here to mimick API response from the backend
-    return Promise.resolve(charities);
+  return fetch("/charity")
+    .then((response) => response.json())
+    .then((entries) => {
+      return entries;
+    });
 }
 
 function getRestaurants() {
-    const restaurants = [{
-            name: "Burger King",
-            description: "Enjoy the best-selling, signature flame-grilled WHOPPER® sandwich as well as other top BK favourites such as the velvety smooth Double Mushroom Swiss, hot and crispy Tendercrisp Chicken, irresistible sides such as Onion Rings, HERSHEY’S® Sundae Pie and more!",
-            address: "Fast Food | VivoCity #02-80",
-        },
-        {
-            name: "Burger King",
-            description: "Enjoy the best-selling, signature flame-grilled WHOPPER® sandwich as well as other top BK favourites such as the velvety smooth Double Mushroom Swiss, hot and crispy Tendercrisp Chicken, irresistible sides such as Onion Rings, HERSHEY’S® Sundae Pie and more!",
-            address: "Fast Food | VivoCity #02-80",
-        },
-        {
-            name: "Burger King",
-            description: "Enjoy the best-selling, signature flame-grilled WHOPPER® sandwich as well as other top BK favourites such as the velvety smooth Double Mushroom Swiss, hot and crispy Tendercrisp Chicken, irresistible sides such as Onion Rings, HERSHEY’S® Sundae Pie and more!",
-            address: "Fast Food | VivoCity #02-80",
-        },
-    ];
-
-    // Using Promise here to mimick API response from the backend
-    return Promise.resolve(restaurants);
+  return fetch("/restaurant")
+    .then((response) => response.json())
+    .then((entries) => {
+      return entries;
+    });
 }
 
 // Add with user's information to signin.html
